Redaction is possible and logs to console after user inputs prompt and redaction filter. Also display resizes as writebox (prompter) increases in size.
